new basic example, renamed previous to comprehensive
