+++ examples/basic/component.tsx
@@ -1,112 +1,17 @@
 import * as React from "react"
-import {
-  dev,
-  FieldSize,
-  StringControl,
-  NumberControl,
-  BooleanControl,
-  SelectControl,
-  ObjectControl,
-  TupleControl,
-  ArrayControl,
-} from "@TMMF/react-dev-toolkit"
+import { dev, NumberControl, StringControl } from "@TMMF/react-dev-toolkit"
 
-const useDevField0 = dev(NumberControl({ name: "Test Title Value 3" }))
+const useDevExampleString = dev(StringControl({ name: "Example String" }))
+const useDevExampleNumber = dev(NumberControl({ name: "Example Number" }))
 
-const useDevField1 = dev(
-  StringControl({
-    name: "Test Title Value 1",
-    // validation: undefined,
-    defaultValue: "test",
-  }),
-)
-
-const useDevField2 = dev(
-  StringControl({
-    name: "Test Title Value 2",
-    size: FieldSize.Large,
-  }),
-)
-
-const useDevField3 = dev(
-  ObjectControl({
-    name: "Test Object Value 4",
-    fields: {
-      "Field 0": StringControl({ name: "Field 0 (Custom Name)" }),
-      "Field 1": NumberControl(),
-      "Field 2": NumberControl(),
-      "Field 3": StringControl(),
-    },
-  }),
-)
-
-const useDevField4 = dev(
-  ArrayControl({ name: "Test Array Value 5", field: StringControl() }),
-)
-
-const useDevField5 = dev(
-  TupleControl({
-    name: "Test Tuple Value 6",
-    fields: [
-      StringControl({ name: "Index 0 (Custom Name)" }),
-      NumberControl(),
-    ] as const,
-  }),
-)
-
-const useDevField6 = dev(BooleanControl({ name: "Test Boolean Value 7" }))
-
-const useDevField7 = dev(
-  SelectControl({
-    name: "Test Boolean Value 7",
-    options: ["Val1", "Val2", "Val3"],
-  }),
-)
-useDevField7.subscribe(console.log)
-useDevField7.subscribeMocked(console.log)
-
-const Component = () => {
-  const value1 = useDevField1("Original Value 1")
-  const value2 = useDevField1("Original Value 2")
-  const value3 = useDevField2("Original Value 3")
-  const value4 = useDevField0(123)
-  const value5 = useDevField3({
-    "Field 0": "test 1",
-    "Field 1": 123,
-    "Field 2": 456,
-    "Field 3": "test 2",
-  })
-  const value6 = useDevField4(["TEST"])
-  const value7 = useDevField5(["TEST", 123])
-  const value8 = useDevField6(false)
-  const value9 = useDevField7("Val1")
-
-  const onClickGET = () => {
-    console.log(
-      useDevField7.getValue(),
-      useDevField7.getValue("Val1"),
-      useDevField7.isMocked(),
-    )
-  }
-
-  const onClickSET = () => {
-    useDevField7.setValue("Val2")
-    useDevField7.setMocked(true)
-  }
+const Component = (props: { exampleString: string; exampleNumber: number }) => {
+  const exampleString = useDevExampleString(props.exampleString)
+  const exampleNumber = useDevExampleNumber(props.exampleNumber)
 
   return (
     <>
-      <p>Value 1a: {value1}</p>
-      <p>Value 1b: {value2}</p>
-      <p>Value 2: {value3}</p>
-      <p>Value 3: {value4}</p>
-      <p>Value 4: {JSON.stringify(value5)}</p>
-      <p>Value 5: {JSON.stringify(value6)}</p>
-      <p>Value 6: {JSON.stringify(value7)}</p>
-      <p>Value 7: {value8?.toString()}</p>
-      <p>Value 8: {value9}</p>
-      <button onClick={onClickGET}>GET VALUE8 IN CONSOLE</button>
-      <button onClick={onClickSET}>SET VALUE8 IN CONSOLE</button>
+      <p>Example String: {exampleString}</p>
+      <p>Example Number: {exampleNumber}</p>
     </>
   )
 }
