+++ examples/comprehensive/index.html
@@ -1,0 +1,8 @@
+<html>
+  <head>
+    <script type="module" src="index.tsx"></script>
+  </head>
+  <body>
+    <div id="root"></div>
+  </body>
+</html>
