+++ package.json
@@ -8,7 +8,8 @@
   },
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
-    "example:basic": "parcel examples/basic/index.html"
+    "example:basic": "parcel examples/basic/index.html",
+    "example:comprehensive": "parcel examples/comprehensive/index.html"
   },
   "repository": {
     "type": "git",
